Deleted the extra comments
So that our code is clean
