+++ User.java
@@ -1,16 +1,15 @@
-import java.sql.Connection; //imports class Connection to establish the connection with mysql database
-import java.sql.PreparedStatement; //imports interface PreparedStatement that helps execute an sql query with parameters
-import java.util.Scanner; //imports class Scanner from java library for interaction with the user
-import java.util.Calendar; //imports class Calendar from java library for user's date of birth
+import java.sql.Connection; 
+import java.sql.PreparedStatement; 
+import java.util.Scanner; 
+import java.util.Calendar; 
 import java.util.ArrayList; 
 
 	public class User {
 		
-	    private String username; //instance variable username that stores the username of the user who wants to sign up
-	    private String password; //instance variable password that stores the password of the user who wants to sign up
-	    private Calendar dateOfBirth; //instance variable dateOfBirth that stores the date of birth of the user who wants to sign up
-	    private String email; //instance variable email that stores the email of the user who wants to sign up
-	    private static ArrayList<String> users = new ArrayList<String> ();
+	    private String username; 
+	    private String password; 
+	    private Calendar dateOfBirth; 
+	    private String email; 
 	    
 	    public User(String name, String pw, Calendar dof, String mail) {
 	    	username = name;
